clickhouse_api/clickhouse_api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
clickhouse-driver==0.2.6
python-dotenv==1.0.0